PayloadSizeBytes -> uint16

+++ ccv/chains/evm/deployment/v1_7_0/sequences/configure_chain_for_lanes.go
@@ -634,7 +634,7 @@ func configureCommitteeVerifierAsSource(
 			AllowlistEnabled:    remoteConfig.AllowlistEnabled,
 			FeeUSDCents:         remoteConfig.FeeUSDCents,
 			GasForVerification:  remoteConfig.GasForVerification,
-			PayloadSizeBytes:    uint16(remoteConfig.PayloadSizeBytes),
+			PayloadSizeBytes:    remoteConfig.PayloadSizeBytes,
 		}
 		currentRemoteReport, err := cldf_ops.ExecuteOperation(b, committee_verifier.GetRemoteChainConfig, chain, contract.FunctionInput[uint64]{
 			ChainSelector: chain.Selector,
